requirements.txt
# Core dependencies
tianshou==2.0.0
gymnasium>=0.29.0
torch>=2.0.0
numpy>=1.21.0
pygame>=2.0.0
matplotlib>=3.5.0
ipython>=8.0.0
jupyter>=1.0.0

# Continuous control environments (optional)
# mujoco>=3.0.0
# gymnasium[mujoco]>=0.29.0
# pybullet>=3.2.6

# Experiment tracking (optional)
# wandb>=0.15.0
# mlflow>=2.0.0

# YAML support for configs
pyyaml>=6.0

# Web UI
streamlit>=1.28.0
tensorboard>=2.0.0

# Development (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# mypy>=1.0.0
# black>=23.0.0
# isort>=5.12.0
# flake8>=6.0.0
# pre-commit>=3.0.0
